fixing cleanup action
